added
ThreadIf::waitForShutdown(int ms) const

which causes the thread to sleep for the specified# of ms, but will wakeup and
return true if shutdown is called. Returns false if the sleep finished w/out
shutdown being called.

shutdown&join are now called in the destructor.  it is now ok to call join multiple
times, nothing happens if there isn't an active thread.