added
ThreadIf::waitForShutdown(int ms) const

which causes the thread to sleep for the specified# of ms, but will wakeup and
return true if shutdown is called. Returns false if the sleep finished w/out
shutdown being called.

shutdown&join are now called in the destructor.  it is now ok to call join multiple
times, nothing happens if there isn't an active thread.


git-svn-id: https://svn.resiprocate.org/rep/resiprocate@828 ddefafc4-47db-0310-ae44-fa13212b10f2