fix contributed by Frederik Bayart
"If I test fdset.readyToRead(conn->getSocket()) immediately after
TcpBaseTransport.cxx:236, this test returns true, although the socket didn't
exist at the time of the select (I'm using solaris 2.8 and gcc 3.2.2). This
has as consequence that on TcpBaseTransport.cxx:133, the test
fdset.readyToRead(currConnection->getSocket()) succeeds and the stack tries
to read even if the connection setup is not yet completed. I didn't verify
whether select sets a bit in the fdset for non-existing sockets, but I
solved the problem by clearing the flag for the socket immediately after
socket creation :"


git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main/sip/resiprocate@2972 ddefafc4-47db-0310-ae44-fa13212b10f2