Code complete for very simple server-side handling of REGISTER requests. This code has not yet been tested in any way except to verify that it compiles, so caveat implementor.
I think I got all this largely correct and consistent
with the rest of the DUM (although it's a bit unclear
what was intended with the registration persistence
class, so I had to kind of make up an API there; I think
it should be very easy to derive an object from what
I currently have for storage to disk, storage to
a relational database, and/or sharing with a mated
pair over a message bus). For testing purposes, I added
a rudimentary persistence class (InMemoryRegistrationDatabase)
that doesn't actually persist data anywhere but the heap.

Also, it's really unclear to me who is responsible
for cleaning up the BaseUsage/NonDialogUsage objects
in general (and the ServerRegistration objects in
particular), so there may be a memory leak in this
at the moment. Once I figure out how these objects
are managed, I'll verify that the ServerRegistration
objects get cleaned up properly, and fix it if
they dont.

Finally, the current interface to the Persistence
manager is a bit rudimentary: it only handles the
URI and expires values. Discarding the display name
portion of contact headers is hardly a problem,
and not keeping around unknown parameters is unlikely
to ever cause issues; however, this code really should
remember q= values. This oversight should be relatively
easy to fix if it becomes important to anyone. In
fact, I'll probably address it myself at some point
in the near future, but my goal for this weekend is
to get something that works, not something that is
100% perfect.