Changes to compat.hxx and udp.h to fix compatibility with Mac OS X headers.
Unfortunately there are 3 very different environment for compiling under OS X that we must support.

Environment 1: command line ./configure & make. In this enviornment none of the native Mac OS X headers are included. Here __APPLE__ is defined but TARGET_OS_MAC is not.

Environment 2: application built with CodeWarrior or XCode using Mac OS 10.3 headers or above. Typical Mac OS X end point developers will have this environment. Mac OS X native headers will be already included by the time resip headers are included. Here __APPLE__ and TARGET_OS_MAC are defined. MAC_OS_X_VERSION_MIN_REQUIRED will be set to the SDK they are targetting.

Environment 3: same as Environment 2 except targgeting 10.2 or less. The socket headers in this SDK are completely different from 10.3 or above. Only Mac developers (such as XTen) maintaining older applications will be using this environment.

In this commit I'm ifdef-ing out typedef for UInt32 which conflicts with the native Mac OS definition. This is a problem for Environment 2 and 3.

Some of these problems would also occur in Environment 1 if we were building under 10.2 or less. BUT: when we detect that we're in Environment 1 (i.e. __APPLE__ defined but TARGET_OS_MAC not defined) we automatically assume that we're in 10.3 or above.

Thus if at a later time we need to have Enviornment 1 compiling on 10.2 or less all we have to do is make sure to include <TargetConditionals.h> at the top of compat.hxx and we'll pop into Environment 3.

I hope this makes things a lot more clear on why compat.hxx has to be the way it is.
thanks sailesh@xten.com



git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@5338 ddefafc4-47db-0310-ae44-fa13212b10f2