Tweak to make work on NetBSD: change shebang line from "#!/usr/bin/perl" to "#!/usr/bin/env perl". This is somewhat more risky, since it searches the current path to find perl; it's also much more portable, since it searches the current path to find perl.
Of course, if you're on a system without perl (or with perl
but not in your path), then there's not much we can do for
you here. There might be some makefile magic to drop in a
stock Makefile.config file in the case that the configure
script fails to generate something.


git-svn-id: https://svn.resiprocate.org/rep/resiprocate@5517 ddefafc4-47db-0310-ae44-fa13212b10f2