Based on a call with Eric and David -
Added a local static guard so that the Log::init
code only _really_ gets run once (or so). Before
this, it was getting run by each compilation unit
which exhausted the number of keys the pthread
library on OS/X (at least) was willing to hand
out. Unfortunately, other things (the gethostbyname
implementation for example) weren't looking to
see if they actually got keys when they asked and
it was all going off the rails.

Thanks to Eric McMurry for chasing this one down.


git-svn-id: https://svn.resiprocate.org/rep/resiprocate@5557 ddefafc4-47db-0310-ae44-fa13212b10f2