added capability to set specific log level for a subsystem.
Use Log::level, not the direct method on the Subsystem. See testLogger for detailed usage.

Also changed Log::Guard to not copy subsytem; it looks like this was resulting in a data copy every log statement. With the new per susbsytem log level, this could result in a lost log mesage, if Subsystem:getLevel is not atomic. Seems worth the risk to avoid the copy.