Remove unused (and non-threadsafe) Timer::mTimerCount/Timer::mId
Previously, all timers were assigned a "unique" (not really, more on that in a 
moment) integer identifier. There is no place in the resip codebase that 
actually uses this identifier in any way. For transaction timers, this 
identifier is in principle unnecessary, since there is more than sufficient 
identifier information present already (the transaction id and timer type). When 
passing a Message* into the timer queue, a unique identifier already exists; the 
Message* itself (if potential use of this Message* bugs you, you can always turn 
it into a handle by applying some sort of transformation to it). This identifier 
is unnecessary in every case I can think of. In addition, the values are 
assigned simply by incrementing a global variable (Timer::mTimerCount), with no 
threadsafety measures whatsoever, so it is not even guaranteed to be unique. 
Because of all this, it has been removed. As a bonus, this saves some memory; 8 
bytes per timer on 64-bit platforms, which adds up to around 3MB when testStack 
steady state has close to 400000 timers in the timer queues at any given point. 
This could be an even larger amount for TUs that schedule lots of long-lifetime 
timers (Timer C, for instance).



git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-TKLC-perf_work@9193 ddefafc4-47db-0310-ae44-fa13212b10f2