Reduce wasted memory in timer queues
Slight refactoring of Timer so that transaction timers and payload timers (ie; 
timers that carry a Message*) are separate classes. Transaction timers no longer 
have an unused Message* member, and payload timers no longer have the unused 
transaction-id, Timer::Type, and duration. This saves a _lot_ of memory for apps 
that use lots of app timers with long lifetimes.



git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-TKLC-perf_work@9196 ddefafc4-47db-0310-ae44-fa13212b10f2