Less wasteful population of Call-IDs
When generating Call-IDs, Helper was computing an md5 hash of the hostname and 
some salt, hex-encoding it, and then Base64 encoding the hex data. We now Base64 
encode the md5 hash directly. This is less computationally expensive, requires 
less memory because the resulting string is half the size, and requires fewer 
bytes on the wire.



git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-TKLC-perf_work@9197 ddefafc4-47db-0310-ae44-fa13212b10f2