Select HashMap implememnation based on libstdc++ version
The current implementation performs the selction based on the
compiler version. When mixing GCC and Clang compiled objects
on RHEL6 this leads to ABI incompatible objects because Clang
pretends to be GCC 4.2 and the system GCC is 4.4.

This patch performs the decision based on the GNU libstdc++
version which is being used. This ensures the same HashMap
implementation across different compilers on a system which
share the same stdlib.

Signed-off-by: Gregor Jasny <Gregor.Jasny@citrix.com>
Reviewed-by: Jupp Mueller <Jupp.Mueller@citrix.com>