Update the Wants and After details for the repro service
Service fails to begin on Fedora 21, this seems to work around the problem.  I started a conversation with Daniel Pocock via the repro-users distribution list and he asked that I provide a pull request so that people can discuss the pros and cons of this particular solution in the pull request.