Update the Wants and After details for the reTurn service
I noticed that the repro service was not starting up as it should do n Fedora 21 for me.  I disucssed this via the repro-users mailing list with Daniel Pocock and he asked that I generate a patch and pull request for both the repro and reTurnServer so that the pros and cons of this particular solution could be discussed in the pull request.