Fix for a buffer management bug in parsing frames from a websocket.
In buffer management/parsing, turns out that if you had a piece of a SIP
packet in the previous buffer (input), and in the current one the remainder
of that packet and another one, it will just copy until the end of input,
instead of just the needed length.

The calculations were also done at the wrong place, they had to be done after
the allocation of the new buffer.

The bug meant that:
- Some packets were getting eaten, resulting in random disconnections and
other dropped events;
- Sometimes repro would crash;
- Or, if compiled with ASAN, it would always crash.

Also, it was more pronounced with TLS.