Removed sources auto-generated by gperf.
This fixes linking errors when gperf 3.1 is used, introduced by
https://github.com/resiprocate/resiprocate/pull/117. Hash function members
use GPERF_SIZE_TYPE for size arguments, which is expanded to size_t and
equivalent to unsigned long on x86-84. However, the old auto-generated
sources committed in git use unsigned int, which causes linking errors
when libresip.so is built.

By removing the sources we force them to be re-generated by the actual
gperf in the system, which resolves the mismatch.