rutil: fix thread race in DnsUtil::getLocalHostName() static initialization
The previous code used only a single gate check before acquiring the lock,
which can lead to the case where multiple threads enter the
initialization block, with subsequent threads re-allocating the buffer of
the localHostName static member, leading to use-after-free memory reads.
This was detected by ASAN.

The fix is to rely on C++11 thread safe initialization guarantees for
local statics.