resip/stack: SipStack, Transport: do not process empty log handlers
The log handlers execution leads to the crash due to null shared
pointers in the list. The protection during adding is required to fix it.

This should fix the recon/unitTest crash where the UserAgent object
could be initialized by the empty SIP message logging handler from the master
profile if the last one does not have it.