Improve logs in DUM for easier tracking error
In ClientInviteSession, moving the log statement above the sendBye(); Currently placing the log after sendBye can lead to confusion, as it makes it unclear why the BYE was sent unless we continue checking later logs.

Typically, when troubleshooting, we focus on the logs which are before the event, so having the log before the sendBye() call provides better context and clarity.