Add support for session refresh using OPTIONS.
When a SIP endpoint does not support timers extension (RFC4028) but
supports OPTIONS request (which it should, as this is one of the basic
requests defined in RFC3261), using OPTIONS instead of re-INVITE to
perform session refresh can be beneficial. OPTIONS is much more
lightweight as it doesn't update SDP, which in some cases may disrupt
media exchange.

Note that OPTIONS can be sent in- or out-of-dialog. In this
implementation we are sending in-dialog OPTIONS to verify that this
dialog is still active.