Add support for iteration over SIP headers.
This refactors the mHeaders and mHeaderIndices containers in SipMessage
into a dedicated KnownHeaders container and exposes this container to
users via the new getRawHeaders() accessor. This allows users to iterate
over the known headers of a SIP message to implement their own logic
dependent on header values.

The fundamental design of the container remains the same, although there
are some changes. In order to implement iterators, the container
elements need to store header type IDs (since, until C++20, iterators'
reference type must be an actual reference, and we want the header type
to be accessible during iteration). We also reuse this field internally
by the container to mark entries as "unused" when a header is removed
from the container. "Unused" elements are not exposed through iterators
or other public accessors and are reused by the container if the same
header is later inserted. Moving this "unused" property into the
container elements allowed to reduce the size of the lookup index
elements to just one byte per header type. And since index elements are
now zero-based, we no longer have to allocate a dummy unused element at
index 0.

Since the container now supports proper iteration over headers, replaced
instances where it was previously emulated via a loop over all known
header types with proper iteration. This should arguably improve
performance on average, since it is unlikely that a SIP message contains
all or even most of header types supported by libresiprocate.

One side effect of this is that the order of headers observed though
iteration (including when SipMessage is serialized) may change. The
tests where this was the case were updated by reordering the headers.